requirement.txt
ipfsApi==
pyPdf==
reportlab==
firebase-admin==

